backport history fix to 2.1 release
